requirements.txt
alabaster==0.7.12
APScheduler==3.6.3
asgiref==3.2.10
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
backtrader==1.9.76.123
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
cssselect==1.1.0
cvxopt @ file:///C:/Users/chuanchao.peng/Downloads/cvxopt-1.2.5-cp39-cp39-win_amd64.whl
cx-Oracle==8.0.1
cycler==0.10.0
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-cors-headers==3.5.0
django-fernet-fields==0.6
django-import-export==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
docutils==0.16
drf-yasg==1.20.0
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
gensim==3.8.3
grpcio==1.33.2
grpcio-tools==1.33.2
h11==0.11.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
incremental==17.5.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver @ file:///C:/Users/chuanchao.peng/Downloads/kiwisolver-1.2.0-cp39-cp39-win_amd64.whl
lxml==4.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib @ file:///C:/Users/chuanchao.peng/Downloads/matplotlib-3.3.2-cp39-cp39-win_amd64.whl
mysqlclient @ file:///C:/Users/chuanchao.peng/Downloads/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
numpy @ file:///C:/Users/chuanchao.peng/Downloads/numpy-1.19.2%2Bmkl-cp39-cp39-win_amd64.whl
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.7
pandas==1.1.3
parso==0.7.1
peewee==3.13.3
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pydash==4.9.0
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
PyYAML==5.3.1
pyzmq==20.0.0
reportlab==3.5.53
requests==2.24.0
ruamel.yaml==0.16.12
scikit-learn @ file:///C:/Users/chuanchao.peng/Downloads/scikit_learn-0.23.2-cp39-cp39-win_amd64.whl
scipy @ file:///C:/Users/chuanchao.peng/Downloads/scipy-1.5.3-cp39-cp39-win_amd64.whl
seaborn==0.11.0
service-identity==18.1.0
six==1.15.0
smart-open==3.0.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.20
sqlparse==0.4.1
tablib==2.0.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted @ file:///C:/Users/chuanchao.peng/Downloads/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
